requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic[email]
passlib[bcrypt]
python-jose
python-multipart